refactor: 자료형 int로 변경

+++ src/main/java/com/example/burnchuck/domain/attendance/model/response/AttendanceMeetingResponse.java
@@ -15,5 +15,5 @@ public class AttendanceMeetingResponse {
     private final LocalDateTime meetingDatetime;
     private final String status;
     private final int maxAttendees;
-    private final Long currentAttendees;
+    private final int currentAttendees;
 }
